Refactor common code to abstract baseclass
